Update post update function

+++ code/rootfs/etc/init.d/postUpdate.sh
@@ -8,9 +8,18 @@
 # 
 # Change History:
 #  VER.   Author         DATE              Change Description
-#  1.0    Qiwei Wu       Dec. 27, 2020     Initial Release
+#  1.0    Qiwei Wu       Mar. 13, 2021     Initial Release
 #*****************************************************************************
 #!/bin/bash
 
-rm -f /mnt/*
-tar -xzvf /tmp/upgrade.tar.gz /mnt/
+if [[ -e /tmp/upgrade ]]; then
+   echo "Updatting"
+   rm -f /mnt/*
+   cd /mnt/
+   tar -xvf /tmp/upgrade
+   echo "rebooting"
+   reboot
+else
+   echo "no update software file"
+fi
+
